finished imu
compatible with simlink, just with temp added

it will probably have another IO/driver type class acting in between
the nnets and the hardware

+++ Software/Fiduino/newStl/Imu.h
@@ -28,8 +28,8 @@ namespace io {
 		// Return 3d acceleration vector
 		TDVect getAccel();
 
-		// Get ambient temperature (0-100)
-		int getTemp();
+		// Get ambient temperature (Celsius)
+		double getTemp();
 	private:
 		void setupCompass();
 		int readSensor(int addrL, int addrH);
